FastAPI/fastapi_supabase/pyproject.toml
[project]
name = "fastapi-supabase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.6",
    "jinja2>=3.1.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "supabase>=2.11.0",
    "uvicorn>=0.34.0",
]
